% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePhases.R
\name{estimatePhases}
\alias{estimatePhases}
\title{Obtain multi-state RACVM estimates}
\usage{
estimatePhases(CP.table, criterion = "BIC", verbose = TRUE)
}
\arguments{
\item{CP.table}{change point table, i.e. output of \code{\link{getCPtable}} which contains a column of change points and a column of selected models.}

\item{criterion}{an information criterion such as BIC, AIC etc.}

\item{verbose}{whether to print a summary of the results}
}
\value{
a named list of phases (numbered by roman numerals) containing the selected model, estimates, and confidence intervals for each parameter.
}
\description{
Takes a set of change points and fits the best model to each phase in an CVM changepoint analysis.
}
\details{
The output of this function is generally passed to the \code{\link{summarizePhases}} function to obtain a tidy summary of the estiamted phases.
}
\seealso{
\code{\link{summarizePhases}}
}
