% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCandidateChangePoints.r
\name{findCandidateChangePoints}
\alias{findCandidateChangePoints}
\title{Find Candidate Change Points}
\usage{
findCandidateChangePoints(windowsweep, clusterwidth = 0)
}
\arguments{
\item{windowsweep}{A windowsweep opject (matrix), output of \code{\link{sweepRACVM}} function}
}
\description{
Find Candidate Change Points
}
