% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucvm.bcpa.functions.r
\name{findSingleBreakPoint}
\alias{findSingleBreakPoint}
\title{findSingleBreakPoint}
\usage{
findSingleBreakPoint(Z, T, plotme = TRUE, method = "sweep", ...)
}
\arguments{
\item{Z}{}

\item{plotme}{}

\item{method}{}

\item{...}{}
}
\description{
findSingleBreakPoint
}
