% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAFfunctions.r
\name{getEVAF}
\alias{getEVAF}
\title{getEVAF}
\usage{
getEVAF(Z, XY = NULL, V = NULL, T, lagmax = max(T)/2)
}
\arguments{
\item{Z}{complex vector of locations}

\item{XY}{two-column matrix of x-y coordinates of locations}

\item{V}{complex vector of velocities}

\item{T}{times of observations}

\item{lagmax}{maximum lag to compute the empirical VAF for}
}
\value{
A data frame with columns: \code{lag} and \code{vaf}
}
\description{
Obtain empirical velocity autocorrelation function.
}
\details{
Requires at least one of complex locations, XY coordinates, or velocties to compute.
}
\examples{
nu <- 2; tau <- 5
ucvm <- simulateUCVM(nu=nu, tau = tau, T.max = 1000, dt = .1)
evaf <- with(ucvm, getEVAF(Z=Z,T=T, lagmax = 20))
plot(evaf)
eta <- 2/sqrt(pi)*nu
curve(eta^2 * exp(-x/tau), add=TRUE, col=2, lwd=2)
}

