% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigma.VZ.r
\name{getSigma.VZ}
\alias{getSigma.VZ}
\alias{getSigma.ZZ}
\title{Obtain 1D complete VV-VZ-ZZ or ZZ covariance matrix}
\usage{
getSigma.VZ(T, nu, tau, t0 = 0)

getSigma.ZZ(T, nu, tau, t0 = 0)
}
\arguments{
\item{T}{vector of times (length n).}

\item{nu}{mean speed.}

\item{tau}{characteristic time scale.}

\item{t0}{initial time defaults to 0}
}
\value{
either a 2n x 2n VV-VZ-ZZ matrix or n x n ZZ matrix.
}
\description{
Takes vector of times T and parameters nu, tau, v0, and returns either complete V-Z covariance matrix (if \code{getSigma.VZ}) or Z-Z portion of the covariance matrix (if \code{getSigma.ZZ}).
}
\examples{
require(gridExtra)
T <- cumsum(rexp(10))
# separated into Sigma_vv, Sigma_vz (assymetric), and Sigma_zz
Sigma <- Matrix(getSigma.VZ(T, nu=2, tau=5, v0=2))
i1 <- image(Sigma[1:10,1:10], main="V-V", colorkey=TRUE, useRaster=TRUE)
i2 <- image(Sigma[11:20,1:10], main="V-Z", colorkey=TRUE, useRaster=TRUE)
i3 <- image(Sigma[11:20,11:20], main="Z-Z", colorkey=TRUE, useRaster=TRUE)
grid.arrange(i1,i2,i3,ncol=3)
# And the complete matrix (log transformed):
image(log(Sigma), colorkey=TRUE, main="log(Sigma)")
}
