% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getV.spline.R
\name{getV.spline}
\alias{getV.spline}
\title{Get splined velocity estimates}
\usage{
getV.spline(Z, T.raw, T.new = NULL, resolution = 0.01, ...)
}
\arguments{
\item{Z}{location data in complex form (X + iY)}

\item{T.raw}{time of observations to feed the spline.}

\item{T.new}{new time vector}

\item{resolution}{of spline, i.e. the fraction of the smallest interval in T.raw.}

\item{...}{further arguments to \code{\link{splinefun}}.  Importantly, the method.}
}
\value{
a list with the estimated velocity (V), as well as the position (Z) and time (T).
}
\description{
Computes the locations and velocities of a cubic spline of a 2D track over time.
}
\examples{
data("Kestrel",package = "smoove")
Z <- with(Kestrel, X+1i*Y)[1:500]
T <- Kestrel$timestamp[1:500]
vsp <- getV.spline(Z,T, resolution = 1e-2, T.new = NULL)
str(vsp)
}
