% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Kestrel}
\alias{Kestrel}
\title{Kestrel}
\format{A data frame with xx rows and n columns
 \describe{
 \item{event.id}{x coordiante}
 \item{visible}{y coordinate}
 \item{Time}{time in POSIXct format}
 }}
\source{
website
}
\usage{
Kestrel
}
\description{
Description of the Kestrel
}
\keyword{datasets}
