% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Lamprey}
\alias{Lamprey}
\title{Lamprey}
\format{A data frame with xx rows and n columns
\describe{
\item{X}{x coordiante}
\item{Y}{y coordinate}
\item{Time}{time in POSIXct format}
}}
\source{
website
}
\usage{
Lamprey
}
\description{
Description of the Lamrey
}
\keyword{datasets}
