% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionRACVM.r
\name{partitionRACVM}
\alias{partitionRACVM}
\title{Obtain multi-state RACVM estimates}
\usage{
partitionRACVM(Z, T, CPs, criterion = "BIC")
}
\arguments{
\item{Z}{locations as a complex vector}

\item{T}{times vector}

\item{CPs}{change points (as indices)}

\item{criterion}{One of \code{"AIC"} or \code{"BIC"} - which information criterion to use for the model selection.}
}
\value{
a named list of phases (numbered by roman numerals) containing the selected model, estimates, and confidence intervals for each parameter.
}
\description{
Takes a set of change points and fits the best model to each phase
}
