% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.track.r
\name{plot.track}
\alias{plot.track}
\title{Plot Track}
\usage{
\method{plot}{track}(track, pch = 19, col = rgb(0, 0, 0, 0.2), cex = NULL,
  type = NULL, ...)
}
\arguments{
\item{track}{a two-column matrix matrix or a complex location vector.}

\item{pch}{default point style (filled circle)}

\item{col}{default color (transluscent grey)}

\item{...}{other arguments to pass to \code{plot}}
}
\description{
Default method for plotting tracks
}
\examples{
data(Bowhead)
bh <- data.frame(X = Bowhead$x.km, Y = Bowhead$y.km)
plot.track(bh)
}

