% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_track.R
\name{plot_track}
\alias{plot_track}
\title{Plot Track}
\usage{
plot_track(
  track,
  pch = 19,
  col = rgb(0, 0, 0, 0.2),
  cex = NULL,
  type = NULL,
  ...
)
}
\arguments{
\item{track}{a two-column matrix matrix or a complex location vector.}

\item{pch}{default point style (filled circle)}

\item{col}{default color (transluscent grey)}

\item{cex}{character expansion}

\item{type}{type of the plot}

\item{...}{other arguments to pass to \code{plot}}
}
\description{
Default method for plotting tracks
}
\examples{
library(smoove)
data(Kestrel)
plot_track(Kestrel[,c('X','Y')])
}
