% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhaseParameter.R
\name{plotPhaseParameter}
\alias{plotPhaseParameter}
\alias{getPhaseParameter}
\title{Plot Estimates From RACVM Partition}
\usage{
plotPhaseParameter(variable, phaselist, cols = 1:length(phaselist),
  label = TRUE, ymax = 1000, ...)
}
\arguments{
\item{variable}{one of: \code{omega}, \code{mu.x}, \code{mu.y}, \code{eta}, \code{tau}}

\item{phaselist}{output of \code{\link{estimatePhases}}, i.e. a list of information on each of the selected phases}

\item{cols}{colors of bars (by default - a rich color palette)}

\item{label}{whether to label the plotted parameter (with the word "tau", "eta", etc.)}

\item{...}{additional arguments to pass to plotting function.  A useful one is \code{log = "y"} for highly skewed data.}
}
\value{
\code{getVariable} returns a data.frame with start time, end time, estimate, low and high 95\% C.I. and selected model for a given parameter, with each row corresponding to an enumerated partition phase.
}
\description{
Plot Estimates From RACVM Partition
}
\details{
\code{getVariable} takes a partitionlist and generates a table indicating the start time, end time, estimate, low and high 95\% C.I. and selected model for a given parameter.

\code{plotVariable}
}
