% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWindowSweep.R
\name{plotWindowSweep}
\alias{plotWindowSweep}
\title{Plot Window Sweep}
\usage{
plotWindowSweep(windowsweep, ...)
}
\arguments{
\item{windowsweep}{output of \code{\link{sweepRACVM}} function}

\item{...}{additional parameters to pass to \code{\link{plot}}}
}
\description{
Plots a relative likelihood plot for the windowsweep
}
