% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.track.r
\name{scan.track}
\alias{scan.track}
\title{Scan Track (enhanced plotting with time)}
\usage{
scan.track(track = NULL, time = NULL, x = NULL, y = NULL, z = NULL,
  set.layout = TRUE, ...)
}
\arguments{
\item{time}{time vector}

\item{z}{optional complex location vector}

\item{...}{options to be passed to plot functions}

\item{{x, y}}{coordinates; can be two separate vectors OR a two-column matrix/date-frame.}

\item{{set.layout}}{whether or not to provide the (default) layout of x-y on the left, and t-x, t-y to the right.}
}
\description{
Plotting x-y, time-x, time-y
}
\examples{
data(Bowhead)
par(bty="l", mar = c(0,2,0,2), oma=c(4,0,4,0), xpd=NA) 
 with(Bowhead, scan.track(time = DateTime, x = x.km, y = y.km))
}
