% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_track.r
\name{scan_track}
\alias{scan_track}
\title{Scan Track (enhanced plotting with time)}
\usage{
scan_track(track = NULL, time = NULL, x = NULL, y = NULL, z = NULL,
  set.layout = TRUE, ...)
}
\arguments{
\item{track}{on object of class track containing locations and times}

\item{time}{time vector}

\item{x}{x coordinates}

\item{y}{y coordinates}

\item{z}{optional complex location vector}

\item{set.layout}{whether to automatically change the layout}

\item{...}{options to be passed to plot functions}
}
\description{
Plotting x-y, time-x, time-y
}
\examples{
data(Bowhead)
par(bty="l", mar = c(0,2,0,2), oma=c(4,0,4,0), xpd=NA) 
with(Bowhead, scan_track(time = DateTime, x = x.km, y = y.km))
}
