% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simSweep}
\alias{simSweep}
\title{simSweep}
\format{A matrix with 400 rows and 24 columns with the following attributes:
 \describe{
 \item{time}{a vector of times}
 \item{Z}{a complex vector of locations}
 
 }}
\source{
website
}
\usage{
simSweep
}
\description{
Description
}
\keyword{datasets}
