% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRACVM.r
\name{simulateRACVM}
\alias{simulateRACVM}
\title{Rotational/Advective Correlated velocity movement}
\usage{
simulateRACVM(tau = 1, eta = 1, omega = 0, mu = 0, v0 = NULL,
  Tmax = 100, dt = 0.01)
}
\arguments{
\item{tau}{characteristic time scale of autocorrelation}

\item{eta}{random rms speed parameter. In the case where \eqn{\omega = 0}, the mean square speed of movement is \eqn{E(|V|^2) = \eta^2 + \mu^2}.  In the case were \eqn{mu = \omega = 0}, the mean tangential speed (elsewhere: \eqn{\nu}) is given by\eqn{ E(|V|) = \eta \sqrt(\pi)/2}.}

\item{omega}{rotational velocity.}

\item{mu}{advective velocity (can be complex value)}

\item{v0}{initial velocity}

\item{Tmax}{max time}

\item{dt}{time interval of simulation}
}
\value{
a list with the following elements \describe{\item{T}{the time vector} \item{V}{the (complex) vector of velocities} \item{Z}{the (complex) vector of locations} \item{XY}{a two-column matrix of x-y coordinates of locations} \item{parameters}{a named list of model parameters.}}
}
\description{
Simulates 2D correlated velocity movement model with rotation and advection
}
\details{
This function simulates a RoAdCVM explicitly using the governing OU equation, using a discretization in which dt << tau.  The model is written
}
\examples{
racvm1 <- simulateRACVM(tau = 4, eta = 1, omega = 2, mu = 0, Tmax = 100, dt = .01)
racvm2 <- simulateRACVM(tau = 4, eta = 1, omega = 0, mu = 1, Tmax = 100, dt = .01)
racvm3 <- simulateRACVM(tau = 4, eta = 1, omega = 2, mu = 1, Tmax = 100, dt = .01)

layout(rbind(c(1,2),c(1,3)))
sapply(list(racvm1, racvm2, racvm3), 
       function(x) with(x, plot(Z, asp=1, type="l", main = paste0("RACVM(", paste(unlist(parameters), collapse = ", "),")"))))


racvm1 <- simulateRACVM(tau = 3, eta = 32, omega = .5, mu = 0, Tmax = 10, dt = .1)
with(racvm1, scan.track(x = Re(Z), y = Im(Z), time = T))


# Compare with simulateRACVM2

racvm1b <- simulateRACVM2(tau = 4, eta = 1, omega = 2, mu = 0, Tmax = 100, dt = .01)
racvm2b <- simulateRACVM2(tau = 4, eta = 1, omega = 0, mu = 1, Tmax = 100, dt = .01)
racvm3b <- simulateRACVM2(tau = 4, eta = 1, omega = 2, mu = 1, Tmax = 100, dt = .01)

racvm1 <- simulateRACVM(tau = 4, eta = 1, omega = 1, mu = 0, Tmax = 10, dt = .01)
racvm1b <- simulateRACVM2(tau = 4, eta = 1, omega = 1, mu = 0, Tmax = 10, dt = .01)
with(racvm1, scan.track(x = Re(Z), y = Im(Z), time = T))
with(racvm1b, scan.track(x = Re(Z), y = Im(Z), time = T))
}
\seealso{
\code{\link{estimateRACVM}}
}
