\name{cvm-package}
\alias{cvm-package}
\alias{cvm}
\docType{package}
\title{
Corellated velocity movement modeling
}
\description{
Estimates correlated velocity movement models (CVM's). 
}
\details{
\tabular{ll}{
Package: \tab cvm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-04\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Elie Gurarie

egurarie@umd.edu
}
\references{
A fun paper is coming out soon. 
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
