% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripModels.R
\name{strip.model}
\alias{strip.model}
\title{A function to strip model fits of all but coefficient estimates}
\usage{
strip.model(model)
}
\arguments{
\item{model}{A fit object to be stripped}
}
\description{
This function was taken from muschellij2/ms_seg, it strips fit objects from R of all but coefficient estimates
}
\keyword{fit}
