% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePhases.R
\name{summarizePhases}
\alias{summarizePhases}
\title{Summarize phases}
\usage{
summarizePhases(phaselist)
}
\arguments{
\item{phaselist}{a phase list (output of \code{\link{estimatePhases}})}
}
\value{
table of start and end times and point estimates of relevant parameters for each phase
}
\description{
Takes a phase list and obtains a tidy table of point estimates of relevant parameters,
}
