% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepRACVM.R
\name{sweepRACVM}
\alias{sweepRACVM}
\title{Sweep RACVM}
\usage{
sweepRACVM(XY = NULL, Z = NULL, T, windowsize, windowstep, model = "UCVM",
  progress = TRUE, ...)
}
\arguments{
\item{XY}{two column location data (if Z is not provided)}

\item{Z}{complex location data (if XY is not provided)}

\item{T}{time vector}

\item{{model}}{model to fit for the change point sweep - typically the most complex model in the candidate model set.}

\item{{windowsize}}{size of window to scan}

\item{{windowstep}}{step by which the window advances}

\item{{progress}}{whether or not to show a progress bar}

\item{{...}}{additional parameters to pass to the \code{\link{estimateRACVM}} function}
}
\description{
Moves a window, computes likelihoods for all change points, and fills out a likelihood matrix.
}
