% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCP.R
\name{testCP}
\alias{testCP}
\title{Test RACVM change point}
\usage{
testCP(Z, T, cp, start, end, modelset = "all", criterion = "BIC",
  spline = FALSE)
}
\arguments{
\item{Z}{location of data}

\item{cp}{change point}

\item{start}{begining timoe of segment}

\item{end}{end time of segment to analyze}

\item{modelset}{set of models to compare (combination of UCVM, ACVM, RCVM, RACVM, or \code{all}, which includes all of them)}

\item{criterion}{}

\item{spline}{}
}
\description{
Identifies appropriate models on either side of a change point
}
