% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCandidateChangePoints.r
\name{findCandidateChangePoints}
\alias{findCandidateChangePoints}
\title{findCandidateChangePoints}
\usage{
findCandidateChangePoints(windowsweep, clusterwidth = 0)
}
\arguments{
\item{windowsweep}{}

\item{clusterwidth}{}
}
\description{
findCandidateChangePoints
}
