% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCandidateChangePoints.r
\name{findCandidateChangePoints}
\alias{findCandidateChangePoints}
\title{Find Candidate Change Points}
\usage{
findCandidateChangePoints(windowsweep, clusterwidth = 0, verbose = TRUE)
}
\arguments{
\item{windowsweep}{A windowsweep opject (matrix), output of \code{\link{sweepRACVM}} function}

\item{clusterwidth}{A time span within which very close change points are considered a single chagne point.}

\item{verbose}{Whether or not to report the number of change points that are clustered away.}
}
\description{
Find Candidate Change Points
}
\details{
The raw output of the window sweep (\code{\link{sweepRACVM}})
}
\examples{
library(smoove)
library(magrittr)
data(simSweep,package = "smoove")
#The warning lets us know that some of the candidate change points are rather too close (in time) to each other
CP.all <- findCandidateChangePoints(windowsweep = simSweep, clusterwidth = 0)
CP.all \%>\% as.vector
CP.clustered <- findCandidateChangePoints(windowsweep = simSweep, clusterwidth = 4)
CP.clustered \%>\% as.vector
}
