% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getS.r
\name{getcov.zz}
\alias{getcov.zz}
\title{Obtain 1D complete S.zz matrix}
\usage{
getcov.zz(T, tau, v0=nu)
getS.ZZ(T, tau, v0=nu)
}
\arguments{
\item{tau}{characteristic time scale.}

\item{T}{vector of times (length n).}
}
\value{
either a 2n x 2n VV-VZ-ZZ matrix or n x n ZZ matrix.
}
\description{
Takes vector of times T and parameters nu, tau, v0, and returns the zz portion of the covariance matrix.
}
\examples{
require(gridExtra)
T <- cumsum(rexp(10))
# separated into Sigma_vv, Sigma_vz (assymetric), and Sigma_zz
Sigma <- Matrix(getSigma.VZ(T, nu=2, tau=5, v0=2))
i1 <- image(Sigma[1:10,1:10], main="V-V", colorkey=TRUE, useRaster=TRUE)
i2 <- image(Sigma[11:20,1:10], main="V-Z", colorkey=TRUE, useRaster=TRUE)
i3 <- image(Sigma[11:20,11:20], main="Z-Z", colorkey=TRUE, useRaster=TRUE)
grid.arrange(i1,i2,i3,ncol=3)
# And the complete matrix (log transformed):
image(log(Sigma), colorkey=TRUE, main="log(Sigma)")
}
