% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhaseParameter.R
\name{getPhaseParameter}
\alias{getPhaseParameter}
\title{Obtain and Plot Estimates From RACVM Partition}
\usage{
getPhaseParameter(variable, phaselist)
}
\arguments{
\item{variable}{one of: \code{omega}, \code{mu.x}, \code{mu.y}, \code{eta}, \code{tau}}

\item{phaselist}{output of \code{\link{estimatePhases}}, i.e. a list of information on each of the selected phases}

\item{...}{additional arguments to pass to plotting function.}
}
\value{
\code{getPhaseParameter} returns a data.frame with start time, end time, estimate, low and high 95\% C.I. and selected model for a given parameter, with each row corresponding to an enumerated partition phase.
}
\description{
Obtain and Plot Estimates From RACVM Partition
}
\details{
\code{getVariable} takes a phase list  and generates a table indicating the start time, end time, estimate, low and high 95\% C.I. and selected model for a given parameter.

\code{plotVariable}
}
\examples{
library(smoove)
library(magrittr)
data(simSweep)

simCP.table <- simSweep \%>\%
  findCandidateChangePoints(clusterwidth = 4, verbose = FALSE) \%>\%
  getCPtable(modelset = c("UCVM", "ACVM"), criterion = "AIC")
simPhaselist <- estimatePhases(simCP.table)

getPhaseParameter("omega",simPhaselist)
}
