% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhases.R
\name{getPhases}
\alias{getPhases}
\title{Obtain multi-state RACVM estimates}
\usage{
getPhases(CP.table, Z, T, criterion = "BIC", verbose = TRUE)
}
\arguments{
\item{CP.table}{change point table, i.e. output of \code{\link{getCPtable}} which contains a column of change points and a column of selected models.}

\item{Z}{locations}

\item{T}{time vector}

\item{criterion}{the criterion to use; defaults to BIC}

\item{verbose}{whether to print a summary of the results}
}
\value{
a named list of phases (numbered by roman numerals) containing the selected model, estimates, and confidence intervals for each parameter.
}
\description{
Takes a set of change points and fits the best model to each phase
}
