% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getV.spline.r
\name{getV.spline}
\alias{getV.spline}
\title{Get splined velocity estimates}
\usage{
getV.spline(Z, T.raw, T.new = NULL, resolution = 0.01, ...)
}
\arguments{
\item{Z}{location data in complex form (X + iY)}

\item{T.raw}{time of observations to feed the spline.}

\item{resolution}{of spline, i.e. the fraction of the smallest interval in T.raw.}

\item{...}{further arguments to \code{\link{splinefun}}.  Importantly, the method.}

\item{T.spline}{times for which the spline is requested.  If NULL (the default) the function returns the velocity at midpoints between consecutive T.raw.}
}
\value{
a list with the estimated velocity (V), as well as the position (Z) and time (T).
}
\description{
Computes the locations and velocities of a cubic spline of a 2D track over time.
}
