% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Kestrel}
\alias{Kestrel}
\title{Kestrel}
\format{A data frame with 7829 rows and 13 columns.
 \describe{
 \item{timestamp}{timestamp}
 \item{longitude}{longitude}
 \item{latitude}{latitude}
 \item{X}{X in UTM coordinate system}
 \item{Y}{Y in UTM coordinate system}

 }}
\source{
https://www.datarepository.movebank.org/handle/10255/move.487
}
\usage{
Kestrel
}
\description{
Movement data from a single kestrel (\emph{Falco naumanni}) in southern Spain. 
See references for data source and vignette for example of analysis.
}
\references{
Hernandez-Pliego J, Rodriguez C, Bustamante J (2015) Why do kestrels soar? PLOS ONE. 10(12): e0145402. doi:10.1371/journal.pone.0145402

Hernandez-Pliego J, Rodriguez C, Bustamante J (2015) Data from: Why do kestrels soar? Movebank Data Repository. doi:10.5441/001/1.sj8t3r11
}
\keyword{datasets}
