% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Kestrel}
\alias{Kestrel}
\title{Kestrel}
\format{A data frame with xx rows and n columns
 \describe{
 \item{event.id}{event.id}
 \item{visible}{visible}
 \item{timestamp}{timestamp}
 \item{longitude}{longitude}
 \item{latitude}{latitude}
 \item{manually.marked.outlier}{manually.marked.outlier}
 \item{sensor.type}{sensor.type}
 \item{individual.taxon.canonical.name}{individual.taxon.canonical.name}
 \item{tag.local.identifier}{tag.local.identifier}
 \item{ID}{ID}
 \item{study.name}{study.name}
 \item{X}{X in UTM coordinate system}
 \item{Y}{Y in UTM coordinate system}
 
 }}
\source{
website
}
\usage{
Kestrel
}
\description{
Description of the Kestrel
}
\keyword{datasets}
