% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhaseList.R
\name{plotPhaseList}
\alias{plotPhaseList}
\title{Plot Phase Partitioning of RACVM analysis}
\usage{
plotPhaseList(
  phaselist,
  cols = gplots::rich.colors(length(phaselist)),
  plot.parameters = TRUE,
  parameters = NULL,
  plot.legend = TRUE,
  legend.where = "bottomright",
  layout = c("horizontal", "vertical")
)
}
\arguments{
\item{phaselist}{Complete "smoove" partitioning output - returned by \link{estimatePhases}}

\item{cols}{colors for phases (by default uses the `rich.colors` palette from `gplots`)}

\item{plot.parameters}{whether the parameters are plotted}

\item{parameters}{which parameters to plot (by default - ALL of the estimated parameters)}

\item{plot.legend}{whether to plot a legend}

\item{legend.where}{location of legend}

\item{layout}{"horizontal" (default) or "vertical"- as preferred (partial string matching accepted)}
}
\description{
Plot Phase Partitioning of RACVM analysis
}
\examples{
library(smoove)

# load data
data(simSweep, package="smoove")

# perform analysis
simCP.table <- simSweep |> 
  findCandidateChangePoints(clusterwidth = 4, verbose = FALSE) |> 
  getCPtable(modelset = c("UCVM", "ACVM"), criterion = "AIC")

# obtain phases
simPhaseList <- estimatePhases(simCP.table)

#plot phases
plotPhaseList(simPhaseList)
}
