% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_track.R
\name{scan_track}
\alias{scan_track}
\title{Scan Track (enhanced plotting with time)}
\usage{
scan_track(
  track = NULL,
  time = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  set.layout = TRUE,
  ...
)
}
\arguments{
\item{track}{on object of class track containing locations and times}

\item{time}{time vector}

\item{x}{x coordinates}

\item{y}{y coordinates}

\item{z}{optional complex location vector}

\item{set.layout}{whether to automatically change the layout}

\item{...}{options to be passed to plot functions}
}
\description{
Plotting x-y, time-x, time-y
}
\examples{
library(smoove)
data(Kestrel, package='smoove')
par(bty="l", mar = c(0,2,0,2), oma=c(4,0,4,0), xpd=NA) 
with(Kestrel, scan_track(time = timestamp, x = X, y = Y))
}
