% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simSweep}
\alias{simSweep}
\title{simSweep}
\format{A matrix with 400 rows and 24 columns with the following notable attributes:
 \describe{
 \item{time}{a vector of times}
 \item{Z}{a complex vector of locations}

 }}
\usage{
simSweep
}
\description{
simSweep is a simulated dataset that could be generate by calling \code{\link{simulateRACVM}} iteratively and then calling \code{\link{sweepRACVM}}. See the provided example.
}
\examples{
if(interactive() && FALSE){
taus <- c(3, 3, 1)
mus <- c(2, 0, 0)
etas <- c(2, 1, 1)
durations <- c(40,60,100)

Z.raw <- 0
T.raw <- 0
mycvm <- list()

for(i in 1:length(taus)){
  if(i > 1)  v0 <- mycvm$V[length(mycvm)]  else v0 = mus[1]
  mycvm <- simulateRACVM(tau = taus[i], eta = etas[i], mu = mus[i], v0 = v0,
                         Tmax = durations[i], dt = 0.01)
  Z.raw <- c(Z.raw, mycvm$Z + Z.raw[length(Z.raw)])
  T.raw <- c(T.raw, mycvm$T + T.raw[length(T.raw)])
}

multicvm <- data.frame(Z = Z.raw, T = T.raw)[sample(1:length(Z.raw), 400),] \%>\%
            plyr::arrange(T)
#a time consuming line of code
simSweep <- with(multicvm, sweepRACVM(Z = Z, T = T,
                 windowsize = 80, windowstep = 5,
                 model = "ACVM", progress=FALSE))
}
}
\keyword{datasets}
