% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCP.R
\name{testCP}
\alias{testCP}
\title{Test RACVM change point}
\usage{
testCP(Z, T, cp, start, end, modelset = "all", spline = FALSE,
  criterion = "BIC", ...)
}
\arguments{
\item{Z}{location}

\item{T}{time of data}

\item{cp}{change point}

\item{start}{beginning time of segment to analyze}

\item{end}{end time of segment to analyze}

\item{modelset}{set of models to compare (combination of UCVM, ACVM, RCVM, RACVM, or \code{all}, which includes all of them)}

\item{spline}{whether or not to use the spline approximation for the final estimate.}

\item{criterion}{selection criterion - either BIC or AIC (can be upper- or lowercased)}

\item{...}{further params to \code{getFit} internal function}
}
\description{
Identifies appropriate models on either side of a change point
}
\examples{
require(smoove)
# Simulate a single change point

set.seed(101)
ucvm1 <- simulateUCVM(T=cumsum(rexp(100)), nu=2, tau=1, method="exact")
ucvm2 <- simulateUCVM(T=cumsum(rexp(100)), nu=5, tau=10, v0 = ucvm1$V[100], method="exact")
T <- c(ucvm1$T, ucvm1$T[100] + ucvm2$T)
Z <- c(ucvm1$Z, ucvm1$Z[100] + ucvm2$Z)
plot_track(Z)

# testing the true changepoint
testCP(Z, T, 100, 1, 200)
# both AIC and BIC are very much lower for the changepoint model

# testing a changepoint in the middle of the first 100 points 
testCP(Z, T, 50, 1, 100)

}
