% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCP.R
\name{testCP}
\alias{testCP}
\title{Test RACVM change point}
\usage{
testCP(Z, T, cp, start, end, modelset = "all", ...)
}
\arguments{
\item{modelset}{set of models to compare (combination of UCVM, ACVM, RCVM, RACVM, or \code{all}, which includes all of them)}

\item{{Z, T}}{location and time of data}

\item{{cp}}{change point}

\item{{start, end}}{beginning and end time of segment to analyze}

\item{spline}{whether or not to use the spline approximation for the final estimate.}

\item{criterion}{selection criterion - either BIC or AIC (can be upper- or lowercased)}
}
\description{
Identifies appropriate models on either side of a change point
}
